+++ packages/orq-rc/RELEASES.md
@@ -4509,3 +4509,13 @@ Based on:
 - [typescript v3.14.0-rc.51] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.51] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.51 - packages/orq-rc
+
+## 2025-10-23 13:48:14
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.638.1 (2.729.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.52] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.52] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.52 - packages/orq-rc